backend-wildlife/backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
argon2-cffi==23.1.0

# API Documentation
drf-yasg==1.21.7

# File uploads
Pillow==10.1.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
httpx==0.26.0
python-dotenv==1.0.0

# Geospatial & Data Processing
numpy<2
pandas==2.0.3
django-location-field==2.7.3
shapely==2.0.2

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
joblib==1.3.2

# Rate limiting
django-ratelimit==4.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==7.0.0

